requirements.txt
psycopg2-binary==2.8.5

daphne==2.5.0
django==3.1
djangorestframework==3.11.1
django-split-settings==1.0.1
celery==4.4.7
pillow==7.2.0

django-webtest==1.9.7
model_bakery==1.1.1
